+++ contact.html
@@ -26,7 +26,7 @@
     <script defer type="module" src="js/main.js"></script>
   </head>
 
-  <body>
+  <body data-page="contact">
 
     <h1 class="hidden">EA. Contact</h1>
     <div id="sticky-nav-con">
