+++ js/modules/animated-skills.js
@@ -1,15 +1,15 @@
 
-// export function animateSkills() { 
-//     gsap.registerPlugin(SplitText)  
+export function animateSkills() { 
+gsap.registerPlugin(SplitText)  
 
-//     const split = new SplitText('.skills', { type: 'chars' })
+const split = new SplitText('.skills', { type: 'chars' })
 
-//     const typing_text = gsap.timeline()
-//     .from(split.chars, {
-//     duration: .1,
-//     autoAlpha: 0,
-//     stagger: {
-//         each: .1
-//     }
-//     });
-// }
+const typing_text = gsap.timeline()
+.from(split.chars, {
+duration: .1,
+autoAlpha: 0,
+stagger: {
+each: .1
+}
+});
+}
